refactor(persistence): ADR-0035 Stage 2.5 C2 slice 2 — extract kirra-fabric-types
Relocate the pure fabric-plane domain types the persistence layer marshals
(FabricAsset + AssetType/KinematicProfileType/AssetPosture/FabricState, and the
CausalLogEntry forensic-ledger record + CAUSAL_EXPORT_MAX_PAGE) into a new lean
crate, kirra-fabric-types. This clears the C2 coupling in verifier_store::fabric,
which named crate::fabric::asset::FabricAsset and
crate::fabric::causal_log::CausalLogEntry. With the ota_campaign slice, both
verifier_store C2 families are now cleared.

What stays in the verifier crate (genuinely not pure data):
- FabricCausalLog — the hashing/signing facade over the shared VerifierStore;
  src/fabric/causal_log.rs keeps it and re-exports the moved CausalLogEntry /
  CAUSAL_EXPORT_MAX_PAGE.
- The KinematicProfileType -> VehicleKinematicsContract mapping
  (nominal_contract / mrc_contract) — verifier-crate behaviour. Since the enum
  now lives in another crate, Rust's orphan rule forbids the inherent impl, so it
  becomes the KinematicProfileContracts extension trait in fabric::governor. All
  call sites are inside that module (the only callers), so call syntax is
  unchanged.

src/fabric/asset.rs becomes a `pub use kirra_fabric_types::asset::*;` shim, so
every crate::fabric::asset::* / crate::fabric::causal_log::CausalLogEntry path
(store, router, telemetry, governor, service binary, tests) resolves unchanged.
Mirrors the kirra-fleet-types / kirra-ota-campaign idiom.

Behaviour-preserving: content moved verbatim; the 2 asset serde round-trip tests
move with the types. Verified: workspace build + `cargo test --lib` (756, = 758 -
the 2 moved) + fabric lib tests (38) + kirra-fabric-types (2) + power-loss drill
(audit_chain_prefix_on_kill, 3) + fmt + quality guardrails + orphan-cores green.
